refactor(evaluation): Restructure evaluation types and update EvaluationExecutor
- Moved the EvaluationResult and EvaluationReport schemas to `evaluation.types.ts` for better organization and modularity.
- Introduced a new `TestCaseEvaluation` interface in `test-case.types.ts` to encapsulate the results of test case evaluations.
- Updated the EvaluationExecutor class to return `TestCaseEvaluation` instead of `EvaluationResult`, enhancing type clarity.
- Adjusted imports across the codebase to reflect the new structure and ensure consistency.

This refactor improves the maintainability and clarity of the evaluation module by consolidating type definitions and enhancing the structure of evaluation results.
